+++ packages/eslint-next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @istok-dev/eslint-next
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @istok-dev/eslint-react@1.0.1
+
 ## 1.0.0
 
 ### Major Changes